WIP; update tokenDB
