+++ src/main/scala/com/ing/wbaa/gargoyle/sts/Server.scala
@@ -4,11 +4,11 @@ import akka.actor.ActorSystem
 import com.ing.wbaa.gargoyle.sts.config._
 import com.ing.wbaa.gargoyle.sts.keycloak.KeycloakTokenVerifier
 import com.ing.wbaa.gargoyle.sts.service.UserTokenDbService
-import com.ing.wbaa.gargoyle.sts.service.db.{ MariaDb, TokenDb }
+import com.ing.wbaa.gargoyle.sts.service.db.MariaDb
 import com.ing.wbaa.gargoyle.sts.service.db.dao.{ STSTokenDAO, STSUserDAO }
 
 object Server extends App {
-  new GargoyleStsService with KeycloakTokenVerifier with UserTokenDbService with TokenDb with STSUserDAO with STSTokenDAO with MariaDb {
+  new GargoyleStsService with KeycloakTokenVerifier with UserTokenDbService with STSUserDAO with STSTokenDAO with MariaDb {
     override implicit lazy val system: ActorSystem = ActorSystem.create("gargoyle-sts")
 
     override protected[this] def httpSettings: GargoyleHttpSettings = GargoyleHttpSettings(system)
